refactor(frontend): move useUser hook
