+++ scoutos-frontend/src/App.tsx
@@ -2,7 +2,8 @@ import ChatInterface from './components/ChatInterface';
 import MemoryManager from './components/MemoryManager';
 import AuthForm from './components/AuthForm';
 import { Toaster } from 'react-hot-toast';
-import { UserProvider, useUser } from './context/UserContext';
+import { UserProvider } from './context/UserContext';
+import { useUser } from './hooks/useUser';
 import './index.css';
 
 function Main() {
