+++ scoutos-frontend/src/components/MemoryManager.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useUser } from "../context/UserContext";
+import { useUser } from "../hooks/useUser";
 
 const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:8000';
