+++ scoutos-frontend/src/hooks/useUser.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { UserContext } from '../context/UserContext';
+
+export function useUser() {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error('useUser must be used within a UserProvider');
+  }
+  return context;
+}
